pyproject.toml
[build-system]
requires = ['setuptools>=61.0']
build-backend = 'setuptools.build_meta'

[project]
name = "mlsgrid_api"
version = "0.0.5"
authors = [
  { name="captivus", email="366332+captivus@users.noreply.github.com" },
]
description = "Python library for the MLSGrid API"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
dependencies = [
  'requests',
  'python-dotenv',
  'pytest',
  'ratelimit',
  'pytz',
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/captivus/mlsgrid_api"